app/requirements.base.txt
crispy-bootstrap4
django
django_celery_results
django-bootstrap5
django-celery-beat
django-cors-headers
django-cron
django-filter
django-latexify
django-revproxy
django-silk
django-storages[s3]
django-tables2
djangorestframework
dotmap
flower
uvicorn
minio
mozilla-django-oidc
psycopg2-binary
redis
watchdog
